Add : refactoring process
