+++ include/threads/thread.h
@@ -16,7 +16,6 @@
  *     @see
  * https://www.notion.so/jactio/write_handler-233c9595474e804f998de012a4d9a075?source=copy_link#233c9595474e80b8bcd0e4ab9d1fa96c
  */
-#include "threads/synch.h"
 #include "userprog/file_abstract.h"
 #endif
 // ADD/write_handler
